feat: add player symbol in footer
