+++ frontend/src/components/PlayerSymbol/index.ts
@@ -1,0 +1,3 @@
+import PlayerSymbol from './PlayerSymbol'
+
+export default PlayerSymbol
