--- frontend/src/components/PlayerSymbolSmall/index.ts
@@ -1,3 +1,0 @@
-import PlayerSymbolSmall from './PlayerSymbolSmall'
-
-export default PlayerSymbolSmall
